KPMP-3988: var rename and fail when files present

+++ data_management/services/dlu.py
@@ -55,13 +55,23 @@ def dlu_file_dict_to_tuple(dlu_file: dict):
 
 
 def create_dest_directory(dest_path: str):
+    return_value = False
     if not os.path.exists(dest_path):
         logger.info("Destination directory " + dest_path + " does not exist. Creating.")
         os.mkdir(dest_path)
+        return_value = True
     else:
-        logger.info("Destination directory exists. Deleting metadata.json if it exists.")
-        if os.path.exists(os.path.join(dest_path, "metadata.json")):
-            os.remove(os.path.join(dest_path, "metadata.json"))
+        dest_dir_info = DirectoryInfo(dest_path)
+        if dest_dir_info.file_count > 1:
+            logger.error("Potential data files in destination directory.")
+        else:
+            if os.path.exists(os.path.join(dest_path, "metadata.json")):
+                logger.info("Deleting metadata.json.")
+                os.remove(os.path.join(dest_path, "metadata.json"))
+                return_value = True
+            else:
+                logger.error("Unknown file in target directory.")
+    return return_value
 
 
 def calculate_checksum(file_path: str):
@@ -75,7 +85,7 @@ def calculate_checksum(file_path: str):
 class DirectoryInfo:
     def __init__(self, directory_path: str):
         self.dir_contents = os.listdir(directory_path)
-        self.dir_count = 0
+        self.subdir_count = 0
         self.file_count = 0
         self.file_details = []
         self.valid_for_dlu = False
@@ -87,7 +97,7 @@ class DirectoryInfo:
         for item in self.dir_contents:
             full_path = os.path.join(self.directory_path, item)
             if os.path.isdir(full_path):
-                self.dir_count += 1
+                self.subdir_count += 1
             else:
                 self.file_count += 1
                 self.file_details.append({
@@ -100,10 +110,10 @@ class DirectoryInfo:
     def check_if_valid_for_dlu(self):
         directory_not_empty = len(self.dir_contents) != 0
         # Nothing but a subdir
-        if self.dir_count == 1 and self.file_count == 1:
+        if self.subdir_count == 1 and self.file_count == 1:
             self.valid_for_dlu = True
         # Items and no subdirectories
-        elif self.dir_count == 0 and directory_not_empty:
+        elif self.subdir_count == 0 and directory_not_empty:
             self.valid_for_dlu = True
         else:
             self.valid_for_dlu = False
@@ -133,26 +143,29 @@ class DLUFileHandler:
         # Make sure the directory is not empty and does not have more than one subdirectory.
         if source_directory_info.valid_for_dlu:
             # Set the source path to the subdirectory if it has only one and is valid.
-            if source_directory_info.dir_count == 1 and source_directory_info.file_count == 0:
+            if source_directory_info.subdir_count == 1 and source_directory_info.file_count == 0:
                 source_package_directory = os.path.join(source_package_directory,
                                                         source_directory_info.dir_contents[0])
                 source_directory_info = DirectoryInfo(source_package_directory)
                 logger.info(
                     "Found one subdirectory (" + source_package_directory + "). Setting it as the main data directory.")
                 if not source_directory_info.valid_for_dlu \
-                        or source_directory_info.dir_count > 0:
+                        or source_directory_info.subdir_count > 0:
                     logger.error("The subdirectory is not valid or there are too many nested subdirectories.")
-                    os.sys.exit()
+                    return False
 
-            create_dest_directory(dest_package_directory)
-            copy_from_src_to_dest(source_package_directory, dest_package_directory)
-            dir_cmp_obj = filecmp.dircmp(source_package_directory, dest_package_directory)
-            # The files that are in the source but not the destination
-            if len(dir_cmp_obj.left_only) == 0:
-                logger.info("Package " + package_id + " moved successfully.")
-                return True
+            create_success = create_dest_directory(dest_package_directory)
+            if create_success:
+                copy_from_src_to_dest(source_package_directory, dest_package_directory)
+                dir_cmp_obj = filecmp.dircmp(source_package_directory, dest_package_directory)
+                # The files that are in the source but not the destination
+                if len(dir_cmp_obj.left_only) == 0:
+                    logger.info("Package " + package_id + " moved successfully.")
+                    return True
+                else:
+                    logger.error("The following files were not copied: " + dir_cmp_obj.left_only.join(","))
+                    return False
             else:
-                logger.error("The following files were not copied: " + dir_cmp_obj.left_only.join(","))
                 return False
         else:
             logger.error("Directory for package " + package_id + " failed validation.")
